requirements.txt
numpy
gym~=0.18.0
tqdm~=4.58.0
stable-baselines3~=1.1.0
einops~=0.3.0
wandb~=0.11.0
accelerate~=0.5.1
pandas
tensorboard~=2.5.0
scipy
pillow
torchattacks~=3.0.0
torchmetrics~=0.5
easydict
vit-pytorch~=0.26.7
timm
